requirements.txt
streamlit
agno
psycopg[binary]
python-dotenv
openai
sqlalchemy
